Bump frontend version to 2.4.1
