+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "frontend",
-  "version": "2.4.0",
+  "version": "2.4.1",
   "private": true,
   "homepage": "./",
   "scripts": {
