+++ frontend/src/components/Test.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars */
 
 // React
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState, useCallback, useMemo } from 'react';
 
 // @mui/material
 
@@ -10,7 +10,8 @@ import React, { useContext, useEffect, useState } from 'react';
 // Frontend
 import { WebSocketContext } from './WebSocketProvider';
 import { Connecting } from './Connecting';
-import { debug } from '../App';
+// import { debug } from '../App';
+const debug = true;
 
 function Test() {
   // WebSocket context
@@ -21,25 +22,37 @@ function Test() {
   const [plugins, setPlugins] = useState([]);
   const [devices, setDevices] = useState([]);
   const [clusters, setClusters] = useState([]);
-
+  const [cpu, setCpu] = useState({});
+  const [memory, setMemory] = useState({});
+  
   useEffect(() => {
     if(debug) console.log('Test useEffect WebSocketMessage mounting');
     const handleWebSocketMessage = (msg) => {
       /* Test page WebSocketMessage listener */
-      if(debug) console.log('Test received WebSocketMessage:', msg.response);
       if (msg.src === 'Matterbridge' && msg.dst === 'Frontend') {
+        if (msg.method === 'restart_required') {
+          if(debug) console.log('Test received restart_required');
+        }
         if (msg.method === 'refresh_required') {
-          if(debug) console.log('Test received refresh_required and sending api requests');
+          if(debug) console.log('Test received refresh_required');
           sendMessage({ method: "/api/settings", src: "Frontend", dst: "Matterbridge", params: {} });
           sendMessage({ method: "/api/plugins", src: "Frontend", dst: "Matterbridge", params: {} });
           sendMessage({ method: "/api/devices", src: "Frontend", dst: "Matterbridge", params: {} });
+        }
+        if (msg.method === 'memory_update') {
+          if(debug) console.log('Test received memory_update', msg);
+          setMemory(msg.params);
+        }
+        if (msg.method === 'cpu_update') {
+          if(debug) console.log('Test received cpu_update', msg);
+          setCpu(msg.params);
         }
         if (msg.method === '/api/settings' && msg.response) {
           if(debug) console.log('Test received settings:', msg.response);
           setSettings(msg.response);
         }
         if (msg.method === '/api/plugins' && msg.response) {
-          if(debug) console.log('Test received plugins:', msg.response);
+          if(debug) console.log(`Test received ${msg.response.length} plugins:`, msg.response);
           setPlugins(msg.response);
         }
         if (msg.method === '/api/devices' && msg.response) {
@@ -54,6 +67,8 @@ function Test() {
           if(debug) console.log(`Test received ${msg.response.length} clusters:`, msg.response);
           setClusters(msg.response);
         }
+      } else {
+        if(debug) console.log('Test received WebSocketMessage:', msg.method, msg.src, msg.dst, msg.response);
       }
     };
 
@@ -70,7 +85,7 @@ function Test() {
   useEffect(() => {
     if(debug) console.log('Test useEffect online mounting');
     if(online) {
-      if(debug) console.log('Test useEffect online sending api requests');
+      if(debug) console.log('Test useEffect online received online');
       sendMessage({ method: "/api/settings", src: "Frontend", dst: "Matterbridge", params: {} });
       sendMessage({ method: "/api/plugins", src: "Frontend", dst: "Matterbridge", params: {} });
       sendMessage({ method: "/api/devices", src: "Frontend", dst: "Matterbridge", params: {} });
@@ -90,6 +105,8 @@ function Test() {
     <div className="MbfPageDiv" style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
       <img src="matterbridge 64x64.png" alt="Matterbridge Logo" style={{ height: '64px', width: '64px' }} />
       <p>Welcome to the Test page of Matterbridge frontend</p>
+      <p>- - -</p>
+      <p>Cpu usage: {cpu.cpuUsage} Uptime: {memory.systemUptime} Memory: freeMemory {memory.freeMemory} totalMemory {memory.totalMemory} rss {memory.rss} heap {memory.heap}</p>
     </div>
   );
 }
